Refactor: Move autoimporter tests to own module
